Moved "Control.Monad.Interface.*" to "Monad.*"
This is shorter and much nicer and more convenient.

The convention is layers is now that every monad interface lives in "Monad.<name of interface>". We can't use the convention(s) used in the mtl, "Control.Monad.<name of interface>{,.Class}", because realistically, given how widely the mtl is used, there are bound to be packages that will end up depending on both layers and mtl, and we shouldn't make it overly difficult (by requiring PackageImports) for them to be used together. The convention we had used previously, "Control.Monad.Interface.<name of interface>", was a bit too verbose, and it makes layers seem more "complicated" than mtl. "Monad.<name of interface>" is as simple as it gets.

Monad transformers still live in "Control.Monad.Trans.<name of transformer<minus the T>>": this is the convention from transformers, and layers is fully compatible with and complementary to transformers, so it shouldn't change this convention.