Bumped version to 1.0.0
I think we've made enough changes in this release to do a really major version bump. I don't want it to be "1.0", because that implies some sort of completeness or stability. Completeness, yes, but it's experimental as fuck, it's really pushing GHC to its limits, and it doesn't even have any kind of test suite (this is something I want to add eventually). It isn't 1.0 though, it's 1.0.0. I feel that slightly communicates the "newness" and untestedness of this better than "1.0".