Dropped the m parameter from LayerState (and LayerEffects)
I can't for the life of me remember what htis was for. I do recollect reading an obscure blogpost somewhere once with an example of a monad that needed it, but it's not used by any of the monads in transformers and I've never used in any of the code I've written that uses layers. It also is preventing me from writing defaults implementations for stacks of 2 and 3. monad-control's equivalent family has no such parameter. I'm inclined to say that any monad transformer that needs this m should take an additional parameter instead (like I do for @resource@'s @SafeT@).