Uncommented and fixed code for ComposeT that I had broken before
This seemed innocuous at first, but to get this fully working I actually had to add a new method `mapI` to `MonadInnerControl`, because it isn't possible to make `OuterResult i m` a proper `Functor` on older GHCs. With defaults and everything else, this should hopefully be a relatively unobtrusive change.